Modify cart and order controllers with views

+++ E-CommerceProject.Entities/ViewModels/CartViewModel.cs
@@ -2,10 +2,10 @@
 
 namespace E_CommerceProject.Entities.ViewModels
 {
-    public class CartViewModel(List<Cart> cartItems, int cartTotal)
+    public class CartViewModel(List<Cart> cartItems, decimal cartTotal)
     {
         public List<Cart> CartItems { get; } = cartItems;
-        public int CartTotal { get; } = cartTotal;
+        public decimal CartTotal { get; } = cartTotal;
 
     }
 }
